librapp/librapp/requirements.txt
Babel==2.2.0
Cheetah==2.4.4
Django==1.9.1
Jinja2==2.8
Landscape-Client==14.12
MarkupSafe==0.23
MySQL-python==1.2.5
PAM==0.4.2
PyYAML==3.10
Pygments==2.1
Sphinx==1.3.5
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
alabaster==0.7.7
apt-xapian-index==0.45
argparse==1.2.1
chardet==2.0.1
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
decorator==4.0.6
django-cors-headers==1.1.0
django-docs==0.1.2
djangorestframework==3.3.2
docutils==0.12
html5lib==0.999
ipython==4.0.2
ipython-genutils==0.1.0
jsonpatch==1.3
jsonpointer==1.0
oauth==1.0.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
prettytable==0.7.2
ptyprocess==0.5
pyOpenSSL==0.13
pycrypto==2.6.1
pycurl==7.19.3
pyserial==2.6
python-apt==0.9.3.5ubuntu1
python-dateutil==2.4.2
python-debian==0.1.21-nmu2ubuntu2
pytz==2015.7
requests==2.2.1
simplegeneric==0.8.1
six==1.5.2
snowballstemmer==1.2.1
sphinx-rtd-theme==0.1.9
ssh-import-id==3.21
traitlets==4.1.0
urllib3==1.7.1
wheel==0.24.0
wsgiref==0.1.2
zope.interface==4.0.5
